+++ fife/lgb_modelers.py
@@ -99,7 +99,9 @@ class GradientBoostedTreesModeler(survival_modeler.SurvivalModeler):
 
         default_params = {"objective": "binary", "num_iterations": 100}
         if n_trials <= 0:
-            return {time_horizon: default_params for time_horizon in range(self.n_intervals)}
+            return {
+                time_horizon: default_params for time_horizon in range(self.n_intervals)
+            }
         params = {}
         if subset is None:
             subset = ~self.data[self.test_col] & ~self.data[self.predict_col]